[MINOR] Fix warnings, imports, code quality, formatting issues

+++ src/main/java/org/apache/sysds/runtime/instructions/spark/ParameterizedBuiltinSPInstruction.java
@@ -524,8 +524,7 @@ public class ParameterizedBuiltinSPInstruction extends ComputationSPInstruction 
 			MultiColumnEncoder encoder = EncoderFactory
 				.createEncoder(params.get("spec"), colnames, fo.getSchema(), (int) fo.getNumColumns(), meta, embeddings);
 			encoder.updateAllDCEncoders();
-			mcOut.setDimension(mcIn.getRows() - ((omap != null) ? omap.getNumRmRows() : 0),
-				(int)encoder.getNumOutCols());
+			mcOut.setDimension(mcIn.getRows() - ((omap != null) ? omap.getNumRmRows() : 0), encoder.getNumOutCols());
 			Broadcast<MultiColumnEncoder> bmeta = sec.getSparkContext().broadcast(encoder);
 			Broadcast<TfOffsetMap> bomap = (omap != null) ? sec.getSparkContext().broadcast(omap) : null;
